refactor(tui::ui::components::popups): move "QuitPopup" to own file
